requirements.txt
black==23.3.0
pylint
joblib==1.2.0
fastparquet==2023.4.0
hyperopt==0.2.7
mlflow==2.3.1
metaflow>=2.8.1
scikit_learn==1.2.2
seaborn==0.12.2
xgboost==1.7.5
orjson==3.8.1
elasticsearch==8.8.0
sentence_transformers==2.2.2
numpy==1.24.3
pandas==2.0.2
nltk==3.8.1
huggingface-hub==0.15.1
openai==0.27.8
matplotlib==3.7.2
wandb==0.15.5
fastcore==1.5.29
plotly==5.15.0
kaleido==0.2.1
basemap==1.3.7
evidently==0.4.0
metaflow-card-html==1.0.1
tools==0.1.9
setuptools 
wheel
python-dotenv
pathlib